backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Environment Configuration
python-dotenv>=1.0.0

# JWT Authentication
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
